+++ Qtica/enums/size.py
@@ -4,4 +4,4 @@ from enum import IntEnum, auto
 class Sizes(IntEnum):
     maximum = auto()
     minimum = auto()
-    size_hint = auto()
+    hint = auto()