+++ Qtica/tools/painters/__init__.py
@@ -1,0 +1,1 @@
+from .status_edge import PaintStatusEdge