+++ Qtica/widgets/window/__init__.py
@@ -1,3 +1,3 @@
 from .mainwindow import MainWindow
 from .routes import RoutingWindow
-from .frameless import FramelessWindow
+from .frameless import FramelessWindow, FramelessWindowSizeGrip