+++ Qtica/widgets/window/routes.py
@@ -1,3 +1,5 @@
+from PySide6.QtCore import Signal
+from PySide6.QtGui import QCloseEvent, QShowEvent
 from PySide6.QtWidgets import (
     QMainWindow, 
     QWidget, 
@@ -10,7 +12,7 @@ from PySide6.QtWidgets import (
 from ...enums.events import EventTypeVar
 from ...enums.signals import SignalTypeVar
 from ...core.base import WidgetBase
-from typing import Dict
+from typing import Dict, Union
 
 
 class Routes:
@@ -40,12 +42,14 @@ class Routes:
 
 
 class RoutingWindow(WidgetBase, QMainWindow):
+    startup_changed = Signal()
+
     def __init__(self, 
                  uid: str = None, 
                  signals: SignalTypeVar = None,
                  events: EventTypeVar = None, 
                  index: str = "/",
-                 routes: Dict[str, QWidget] = None,
+                 routes: Dict[str, Union[QWidget, QLayout]] = None,
                  stacked_widget: QStackedWidget = None,
                  tool_bar: QToolBar = None,
                  status_bar: QStatusBar = None,
@@ -54,6 +58,7 @@ class RoutingWindow(WidgetBase, QMainWindow):
         QMainWindow.__init__(self)
         super().__init__(uid, signals, events, **kwargs)
 
+        self.__is_startup = False
         stacked_widget = (stacked_widget 
                           if stacked_widget is not None 
                           else QStackedWidget())
@@ -65,34 +70,48 @@ class RoutingWindow(WidgetBase, QMainWindow):
         self.setUpdatesEnabled(True)
         self.setDocumentMode(True)
 
-        self._set_routes(routes)
-        self._set_tool_bar(tool_bar)
-        self._set_status_bar(status_bar)
-        self._set_dock_widget(dock_widget)
+        if routes is not None:
+            self._set_routes(routes)
+
+        if tool_bar is not None:
+            self._set_tool_bar(tool_bar)
+    
+        if status_bar is not None:
+            self._set_status_bar(status_bar)
+
+        if dock_widget is not None:
+            self._set_dock_widget(dock_widget)
+
+    def closeEvent(self, event: QCloseEvent) -> None:
+        self.__is_startup = False
+        return super().closeEvent(event)
+
+    def showEvent(self, event: QShowEvent) -> None:
+        if not self.__is_startup:
+            self.__is_startup = True
+            self.startup_changed.emit()
+
+        return super().showEvent(event)
 
     def _set_routes(self, routes: Dict[str, QWidget]) -> None:
-        if routes is not None:
-            for route, child in routes.items():
-                if isinstance(child, QLayout):
-                    widget = QWidget(self)
-                    widget.setLayout(child)
-                    self._routes.add(route, widget)
-                else:
-                    self._routes.add(route, child)
+        for route, child in routes.items():
+            if isinstance(child, QLayout):
+                widget = QWidget(self)
+                widget.setLayout(child)
+                self._routes.add(route, widget)
+            else:
+                self._routes.add(route, child)
 
-            self.setCentralWidget(self._routes.stackedWidget())
+        self.setCentralWidget(self._routes.stackedWidget())
 
     def _set_tool_bar(self, tool_bar: QToolBar) -> None:
-        if tool_bar is not None:
-            return self.addToolBar(tool_bar)
+        self.addToolBar(tool_bar)
 
     def _set_status_bar(self, status_bar: QStatusBar) -> None:
-        if status_bar is not None:
-            return self.setStatusBar(status_bar)
+        self.setStatusBar(status_bar)
 
     def _set_dock_widget(self, dock_widget: QDockWidget) -> None:
-        if dock_widget is not None:
-            self.addDockWidget(dock_widget)
+        self.addDockWidget(dock_widget)
 
     def push(self, route: str) -> None:
         self.route.push(route)