+++ examples/qtica_path.py
@@ -1,8 +1,8 @@
 import sys
 import os
 
-# paths to mifast package
-mifast_pkg = os.path.join(os.path.dirname(__file__), "..")
+# paths to Qtica package
+Qtica_pkg = os.path.join(os.path.dirname(__file__), "..")
 
 # append paths to system environment
-sys.path.append(mifast_pkg)
+sys.path.append(Qtica_pkg)